+++ packages/terra-form-select/CHANGELOG.md
@@ -3,6 +3,9 @@ ChangeLog
 
 Unreleased
 ----------
+### Changed
+* Update tests to use `Terra.describeViewports` for setting viewports and use `Terra.it.validatesElement`
+
 ### Fixed
 * Issue where input would remain focused after clicking away from select menu after enter had been used to select an item in `multiselect` and `tag` variants.